+++ src/Downloader/Http/DownloadData.h
@@ -3,6 +3,7 @@
 #pragma once
 
 #include <chrono>
+#include <cstdint>
 #include <memory>
 #include <optional>
 #include <string>
@@ -19,8 +20,8 @@ class CurlWrapper;
 class DownloadDataPack
 {
 public:
-	int size = 0;
-	int progress = 0;
+	uint64_t size = 0;
+	uint64_t progress = 0;
 };
 
 class DownloadData
@@ -31,7 +32,7 @@ public:
 	std::string mirror;                  // mirror used
 	IDownload* download;
 	DownloadDataPack* data_pack = nullptr;
-	int approx_size = 0;  // Either approx or real size from the IDownload.
+	uint64_t approx_size = 0;  // Either approx or real size from the IDownload.
 	int retry_num = 0;
 	std::chrono::seconds retry_after_from_server{0};
 	std::chrono::steady_clock::time_point next_retry;
